requirements.txt
# Medical Image Classification Requirements
# H200 GPU Server Optimized

# Core ML frameworks - H200 optimized versions
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# MONAI framework
monai[all]>=1.4.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0

# Transformers for text processing
transformers>=4.35.0
tokenizers>=0.15.0
datasets>=2.14.0

# Machine learning utilities
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Progress bars and utilities
tqdm>=4.66.0
wandb>=0.16.0  # For experiment tracking

# Image processing
opencv-python>=4.8.0
scipy>=1.11.0

# Performance optimization for H200
numba>=0.58.0
nvidia-ml-py3>=7.352.0

# Logging and monitoring
tensorboard>=2.15.0

# Additional utilities
argparse
pathlib
json5
pyyaml>=6.0.0
psutil>=5.9.0

# Development and debugging
jupyter>=1.0.0
ipykernel>=6.25.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0